Stabilization Part 4. Config's refactor. * Deleting ExternalConfig->merging. * Ready for future AutoIPConfig feature. * Events/Mods split. * Major cleanup. * Corrected Typos. * Added missing config's. * Deleted unusable config's. * Deleted Automatic IP will be implemented later an intelligent one. * Split Network config's. * Minor fixes
Builders:
* Cleanup.
* Added missing build doc's.
* Removed useless code.

General:
* Useless Files Delete.
* Utilities: comments cleanup (about 460kb codes).
* Typos.
* Removed double gpl license.
* Updated pack's information.
* Organized some java files.
* Stats blow config's bytes->int.
* Debug launchers (still need fixes) updates.
* Datapack cleanup from older revisions.
* Removed old unused GM Classic menu.

Remember to backup your workspace before you do anything.

+++ L2jHellasC/java/com/l2jhellas/gameserver/handler/skillhandlers/Blow.java
@@ -61,7 +61,7 @@ public class Blow implements ISkillHandler
 			// Check firstly if target dodges skill
 			final boolean skillIsEvaded = Formulas.calcPhysicalSkillEvasion(target, skill);
 
-			byte _successChance = Config.SIDE_BLOW_SUCCESS;
+			int _successChance = Config.SIDE_BLOW_SUCCESS;
 
 			if (activeChar.isBehindTarget())
 				_successChance = Config.BACK_BLOW_SUCCESS;
